backend/requirements.txt
Django
psycopg2
django-environ
djangorestframework
django-filter
django-cors-headers
